requirements.txt
bcrypt==4.3.0
fastapi==0.115.9
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.72
langchain-google-genai==2.1.8
psycopg2-binary==2.9.11
pydantic==2.11.5
python-dotenv==1.0.1
uvicorn==0.34.0
